fanxfan: 完善图片验证相关 全局异常处理 密码错误次数限制

+++ exam-entity-jpa/src/main/java/top/fanxfan/jpa/base/service/impl/CaptchaServiceImpl.java
@@ -8,6 +8,7 @@ import jakarta.annotation.Resource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import top.fanxfan.core.captcha.CaptchaProperties;
 import top.fanxfan.core.exception.ServiceException;
 import top.fanxfan.core.tools.RedisUtils;
 import top.fanxfan.jpa.base.service.CaptchaService;
@@ -15,6 +16,7 @@ import top.fanxfan.jpa.base.service.CaptchaService;
 import java.time.Duration;
 import java.util.Map;
 
+import static top.fanxfan.jpa.base.constants.BaseErrorConstants.VERIIFY_CODE_ERROR;
 import static top.fanxfan.jpa.base.constants.BaseRedisKeyConstants.CAPTCHA_KEY;
 
 /**
@@ -28,25 +30,34 @@ public class CaptchaServiceImpl implements CaptchaService {
     @Resource(name = "captcha")
     private final AbstractCaptcha captcha;
 
+    @Resource
+    private CaptchaProperties captchaProperties;
+
 
     @Override
-    public Map<String, String> create() {
+    public Map<String, Object> create() {
+        if (Boolean.FALSE.equals(captchaProperties.getEnable())) {
+            return Map.of("enableCaptcha", Boolean.FALSE);
+        }
         captcha.createCode();
         String uuid = UUID.fastUUID().toString();
         String code = captcha.getCode();
         log.debug("uuid:{}, code:{}", uuid, code);
         RedisUtils.setCacheObject(formatKey(uuid), code, Duration.ofMinutes(3));
         String imageBase64 = captcha.getImageBase64();
-        return Map.of("uuid", uuid, "imageBase64", imageBase64);
+        return Map.of("uuid", uuid, "imageBase64", imageBase64, "enableCaptcha", Boolean.TRUE);
     }
 
     @Override
     public void verify(String uuid, String code) {
+        if (Boolean.FALSE.equals(captchaProperties.getEnable())) {
+            return;
+        }
         String key = formatKey(uuid);
         String cacheValue = Convert.toStr(RedisUtils.getCacheObject(key), "");
         RedisUtils.deleteObject(key);
         if (!CharSequenceUtil.equals(cacheValue, code)) {
-            throw new ServiceException("验证码错误");
+            throw new ServiceException(VERIIFY_CODE_ERROR);
         }
     }
